Changed front-ends call to API (since API has new response structure) and finished form validation
