+++ src/App.js
@@ -1,6 +1,7 @@
 import { BrowserRouter as Router, Route } from "react-router-dom";
+import "./Styles/App.css";
 
-import { Home, Register, Login, Navbar, UsersOnly, Form } from "./Components/index";
+import { Home, Register, Login, Navbar, UsersOnly } from "./Components/index";
 import { PrivateRoute, PublicRoute } from "./HoCs/index";
 
 function App() {
@@ -8,7 +9,6 @@ function App() {
         <Router>
             <Navbar />
             <Route exact path="/" component={Home} />
-            {/*<PublicRoute exact path="/valid" component={Form} />*/}
             <PublicRoute exact path="/register" component={Register} />
             <PublicRoute exact path="/login" component={Login} />
             <PrivateRoute exact path="/usersonly" component={UsersOnly} />
