+++ src/Components/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 const Home = () => {
-    return <h1>Home Page</h1>;
+    return <h1>Home Page, bro</h1>;
 };
 
 export default Home;
