+++ src/Components/Login.js
@@ -1,59 +1,70 @@
-import React, { useState, useContext } from "react";
+import React, { useState, useRef, useEffect, useContext } from "react";
+import { useForm } from "react-hook-form";
+import { yupResolver } from "@hookform/resolvers/yup";
+
 import AuthService from "../Services/AuthService";
 import { AuthContext } from "../Context/AuthContext";
-import Message from "../Components/Message";
+import Message from "./Message";
+import LoginSchema from "../Models/LoginFormModel";
 
 const Login = (props) => {
-    const [user, setUser] = useState({ email: "", password: "" });
-    const [serverMessage, setServerMessage] = useState(null);
+    const [userMessage, setUserMessage] = useState(null);
     const authContext = useContext(AuthContext);
+    let timerID = useRef(null);
 
-    const onChange = (event) => {
-        setUser({...user, [event.target.name]: event.target.value});
-    }
-    const onSubmit = (event) => {
-        event.preventDefault();
-        AuthService.login(user).then((data) => {
-            console.log(data);
-            const { isAuthenticated, user, serverMessage } = data;
+    useEffect(() => {
+        return () => {
+            clearTimeout(timerID);
+        };
+    }, []);
+
+    const { register, handleSubmit, errors, reset } = useForm({
+        resolver: yupResolver(LoginSchema),
+    });
+
+    const submitLoginForm = (userData) => {
+        AuthService.login(userData).then((serverMessage) => {
+            console.log(serverMessage);
+            const { isAuthenticated, user, msgBody } = serverMessage;
+            setUserMessage(msgBody + ", you will soon be redirected");
             if (isAuthenticated) {
-                authContext.setUser(user);
-                authContext.setIsAuthenticated(isAuthenticated);
-                props.history.push("/");
+                timerID = setTimeout(() => {
+                    authContext.setUser(user);
+                    authContext.setIsAuthenticated(isAuthenticated);
+                }, 2000);
             } else {
-                setServerMessage(serverMessage);
+                reset();
             }
-        })
-    }
+        });
+    };
+
     return (
         <div>
-            <form onSubmit={onSubmit}>
-                <h3>Please sign in</h3>
-                <label htmlFor="email" className="sr-only">
-                    Email:
-                </label>
+            <form onSubmit={handleSubmit(submitLoginForm)}>
                 <input
                     type="text"
                     name="email"
-                    onChange={onChange}
-                    placeholder="Enter email"
+                    placeholder="Email..."
+                    ref={register}
                 />
-                <label htmlFor="password" className="sr-only">
-                    Password
-                </label>
+                {errors.email ? (
+                    <Message message={errors.email.message} />
+                ) : null}
+                <br />
                 <input
-                    type="password"
+                    type="text"
                     name="password"
-                    onChange={onChange}
-                    placeholder="Enter password"
+                    placeholder="Password..."
+                    ref={register}
                 />
-                <button
-                    type="submit"
-                >
-                    Log In
-                </button>
+                {errors.password ? (
+                    <Message message={errors.password.message} />
+                ) : null}
+                <br />
+                <button type="submit">Log In</button>
+                <br />
+                {userMessage ? <Message message={userMessage} /> : null}
             </form>
-            {serverMessage ? <Message message={serverMessage} /> : null }
         </div>
     );
 };
