+++ src/Components/index.js
@@ -1,6 +1,7 @@
 export { default as Home } from "./Home";
-export { default as Register } from "./Register";
-export { default as Login } from "./Login";
+export { default as Register_old } from "./Register_old";
+export { default as Login_old } from "./Login_old";
 export { default as Navbar } from "./Navbar";
 export { default as UsersOnly } from "./UsersOnly";
-export { default as Form } from "./Form";
+export { default as Register } from "./Register";
+export {default as Login } from "./Login";
