+++ src/Styles/App.css
@@ -1,0 +1,4 @@
+body {
+    margin: 1em;
+    font-size: 1.5em;
+}
